bot/requirements.txt
discord.py
beautifulsoup4
requests
Flask
python-dotenv